Python/openai-tracing/pyproject.toml
[tool.poetry]
name = "openai-tracing"
description = "Tracing of openai calls"
version = "0.0.0"
authors = ["Autoblocks Engineering <engineering@autoblocks.ai>"]
license = "MIT"
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.9"
openai = "^1.0.0"
autoblocksai = "^0.0.28"
python-dotenv = "^1.0.0"

[tool.poetry.scripts]
start = "src.main:main"
